Separate user overview analysis and user EDA on user files
